Rework how actions are stored and FbInk usage
This commit splits up actions and recordings into two separate files.
The recordings are still stored using bincode but the actions are now
stored as TOML to make it easy to edit e.g. the sort value or
post-playback delay after an action has been recorded.

FbInk is now just initialized once at startup and shared with an Arc to
see if that helps with screenshots failing after the server has been
running for a while on a Sage.