Rework touch handling for arbitrary input
Chromium generates more touchmove input events so the old method for
deciding if a touch was a tap didn't work. This changes things to use
the distance between touch start and end instead of the number of move
events, and also adds the option of starting arbitrary input mode with a
swipe instead of a longpress.