Tweak optimize_events behaviour
The previous implementation wouldn't skip optimizing long-lasting
actions on an Aura H20. This adds a hacky workaround for that,
and stops trying to optimize all actions by default, in case there are
other situations where it doesn't work. Page turns are now explicitly
optimized instead of relying on the default being true.