Rework arbitrary input feature
Switches to axum's WebSocket feature instead of directly using
tungstenite. Adds some new InputMsg variants to enable the InputManager
to be controlled from elsewhere in the code now that it's not managing
the WebSocket directly. Instead of running once when the server first
starts, (re)start the InputManager when the page turner or remote
control pages load, so that config changes take effect without having to
restart the server. This commit should mitigate if not fix a bug that
was occasionally causing the feature to stop working.